Update 2019.04.05
Over the past week (and still not complete) I have been compiling Creature Stats into an excel sheet. These are simple stats with notes and etc, but with the aim to determine which Main species types and variants  I want to invest time into to create codex entries.

The remainder of my time on this project has been being spent re-reading the core books for 5e to get better understandings of the rules as well as doing researching for Monster Creation.