Update 2019-04-27
Finals season is in, so updates are a bit slower. Still, this is a good sized update (IMO).

Worked on figuring out differences for types of monsters (what monsters may be worth creating stat blocks for), as well as creating a few prototype monsters (see `Dune Spider.png` and `Maguuma Spider.png` under `\Prophecies\Assets`.

Feedback on monster stats is always welcomed and encouraged.