Update 2019-05-22
This update comprises a bit of work I've put in over the past few days.

`Creature Encounters.xlsx` is the sheet I am starting to keep encounters in for each area. It is divided by area, with the idea of showing encounters ala Pokemon style. Two areas are "complete" enough to show, and have been given picture versions. These can be seen in `Locations of Guild Wars Prophecies.md`.

Another small project I just touched on is a concept for quests, `Quest Propmts for Guild Wars.md`. This is very much in a conceptual stage, but is a possibility I'm looking at to determine how to show quests in general. There are a TON of quests in GW however, so I think the end result might be best if curated and chosen a couple to "remix" for each area. Whether they will be in a different location apart from the "locations" is yet to be decided.

Additionally, I've put in work on the `Locations of Guild Wars Prophecies.md`:
+ Ascalon Encounter Table
+ Crystal Desert Encounter Table
- Removed "Fauna" areas in favor of Enounter Tables.
+ Kryta: The Black Curtain, North Kryta Province.

As always, any feedback on any of this is greatly appreciated.